fix issue #13 in edifact encoder
Up to now, the edifact encoder only added a return-to-ascii code if
necessary. Apparently, this causes problems with datamatrix readers
- at least in some cases.

=> change code to _always_ add return-to-ascii code. Costing us an
additional byte in affected cases.

Also had to fix test_encode_known() as it expected the encoder to do
switch to edifact mode even for single-character messages (which is
not optimal anyway).

closes #13