svg viewBox
We have to specify a viewBox in svg so that the renderer knows which
part of the canvas is the image (the canvas has infinite size).

Dropped width and height attributes: I don't think I should impose a
particular size on the datamatrix. It should be rendered as large as the
user wants it.

closes #18