Re-added height and width attributes
Defining the viewBox in the previous commit was the right thing to do.
Still, I am re-adding height and width attributes to the generated svg
to give IPython an indication of how big it should render the code.
Without it, it will be rendered to fill the available space which is not
nice.